FileFlow/backend/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
flask-cors
cryptography
flake8
watchdog
py7zr
pillow
python-magic
aiofiles